Moved tests to correct packages
