+++ src/test/java/com/geronimo/GeronimoDatabaseTest.java
@@ -6,28 +6,21 @@ import org.springframework.test.context.junit4.SpringRunner;
 
 import javax.persistence.EntityManager;
 import javax.persistence.PersistenceContext;
-import javax.persistence.Query;
-
 import java.math.BigInteger;
 
-import static org.junit.Assert.*;
+import static org.hamcrest.MatcherAssert.assertThat;
+import static org.hamcrest.Matchers.is;
 
 
 @RunWith(SpringRunner.class)
 @GeronimoSpringTest
 public class GeronimoDatabaseTest {
 
+    @PersistenceContext
     private EntityManager entityManager;
 
     @Test
     public void testDatabaseConnectionEstablished() {
-        Query nativeQuery = entityManager.createNativeQuery("SELECT 1");
-        assertEquals(nativeQuery.getSingleResult(), BigInteger.valueOf(1L));
+        assertThat(entityManager.createNativeQuery("SELECT 1").getSingleResult(), is(BigInteger.valueOf(1L)));
     }
-
-    @PersistenceContext
-    public void setEntityManager(EntityManager entityManager) {
-        this.entityManager = entityManager;
-    }
-
 }
